move `PlayerRatingChart` class from `com.faforever.client.fx` to `com.faforever.client.chat`
